feat: Add onChange prop to Select component
The code changes in the `Select.tsx` file add the `onChange` prop to the `Select` component. This prop allows for a callback function to be passed in, which will be triggered when the selected option in the dropdown changes. This enhancement improves the functionality of the `Select` component by providing a way to handle user interactions and perform actions based on the selected option.